Move ship_config to models
