+++ src/virtualship/cli/_fetch.py
@@ -38,7 +38,7 @@ def _fetch(path: str | Path, username: str | None, password: str | None) -> None
     be provided on prompt, via command line arguments, or via a YAML config file. Run
     `virtualship fetch` on an expedition for more info.
     """
-    from virtualship.expedition.ship_config import InstrumentType
+    from virtualship.models.ship_config import InstrumentType
 
     if sum([username is None, password is None]) == 1:
         raise ValueError("Both username and password must be provided when using CLI.")
