+++ src/virtualship/expedition/simulate_measurements.py
@@ -13,7 +13,7 @@ from ..instruments.ctd_bgc import simulate_ctd_bgc
 from ..instruments.drifter import simulate_drifters
 from ..instruments.ship_underwater_st import simulate_ship_underwater_st
 from ..instruments.xbt import simulate_xbt
-from .ship_config import ShipConfig
+from ..models.ship_config import ShipConfig
 from .simulate_schedule import MeasurementsToSimulate
 
 if TYPE_CHECKING:
