+++ tests/cli/test_fetch.py
@@ -17,7 +17,7 @@ from virtualship.cli._fetch import (
     hash_to_filename,
 )
 from virtualship.expedition.schedule import Schedule
-from virtualship.expedition.ship_config import ShipConfig
+from virtualship.models.ship_config import ShipConfig
 from virtualship.utils import get_example_config, get_example_schedule
