+++ tests/test_mfp_to_yaml.py
@@ -4,7 +4,7 @@ import pandas as pd
 import pytest
 
 from virtualship.expedition.schedule import Schedule
-from virtualship.expedition.ship_config import InstrumentType
+from virtualship.models.ship_config import InstrumentType
 from virtualship.utils import mfp_to_yaml
